// client/src/js/bindings/AdminUserRow.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Game } from "./Game";
import type { Permission } from "./Permission";

export type AdminUserRow = { gamertag: string, game: Game, banished: boolean, connected: boolean, permissions: Array<Permission>, created_at: bigint, };